rename EphemeralSocket.js to ephemeral-socket.js
