+++ lib/ephemeral-socket.js
