+++ lib/statsd-client.js
@@ -1,4 +1,4 @@
-var EphemeralSocket = require('./EphemeralSocket');
+var EphemeralSocket = require('./ephemeral-socket.js');
 var Messages = require('./messages.js');
 
 function StatsdClientOptions(prefix, isDisabled, dnsResolver) {
